requirements.txt
numpy==1.19.4
requests==2.22.0
beautifulsoup4==4.9.3
fasttext==0.9.2
matplotlib==3.3.3
nltk==3.5
pandas==1.1.4
pdf2image==1.14.0
progressbar33==2.4
pytesseract==0.3.6
scikit_learn==0.23.2
scipy==1.5.4
selenium==3.141.0
tika==1.24
wordcloud==1.8.1
